Create LAB RolloverExample
uint creates Uint8 (simply the name of the unsigned integer, the number 8 has no bearing on the value of the integer)
The increment and decrement functions allow the value of Uint8 (0 by default) to be incremented or decremented once the contract is deployed
Within both the increment and decrement functions we added the unchecked {} line, whereas the integer was unable to be decreased beyond 0 or above 255, now that both functions have no bounds, the unsigned integer will loop back to 0 if increased from 255 or would go 0 to 255 if decreased. Note that this is still not a negative integer and the value can still only range freely from 0 to 255 as opposed to the original state of the function without the unchecked{} line, where the value was stopped at 0 and 255.