Create LAB AddressExample
Here we created an address, a function to set the address, and a function to retrieve the account balance. Pretty simple to deploy and run utilize the contract. Some key points mentioned in the lab notes: The smart contract itself is stored under its own address; the address stored under myAddress can be any address, including the smart contract's address; we can retrieve the balance of any address that could be stored in the myAddress variable as blockchain information is public; the contract can transfer funds from the contract address to another address but not vice versa; transferring erc-20 tokens works fundamentally different from transferring ether.