Create LAB StringExample
Creates the string and sets it's value to 'hello world!'
Also creates a function to set myString to whatever value is desired. Once the conract is deployed, I changed the string's value to 'Torgmeister'
Note that strings in solidity are available from libraries and should typically be avoided due to gas costs. Strings are essentially arrays. Natively, there are no string manipulation functions and it is not even possible to compare strings.