Create LAB SendMoneyExample
Lesson in pdf for solidity course starts around page 100. Deployed this contract 3 times and added to it as we went. First, we used receiveMoney getBalance and withdraw money. Moving forward, we added withdrawMoneyTo, then on the third iteration we added the block.timestamp lines to not allow anyone to withdraw until a minute had passed from the timestamp of the last block. More details can be found on the pdf.